fix: renamed files and enums so that their names match
