+++ test/models/ui/map_details_test.dart
@@ -1,7 +1,7 @@
 import "package:flutter_test/flutter_test.dart";
 import "package:google_maps_flutter/google_maps_flutter.dart";
 import "package:proxima/models/ui/map_details.dart";
-import "package:proxima/views/components/options/map/map_selection_option.dart";
+import "package:proxima/views/components/options/map/map_selection_options.dart";
 
 void main() {
   group("hash and == works", () {
